move analytics code to its own folder
this will segregate from server-ish code (sql, etc) and client, which
might be javascript

Of course server code will probably call analytics.
